move dde2 into dde
