+++ myia/opt/__init__.py
@@ -1,5 +1,5 @@
 """Optimization submodule."""
 
 from .cse import *  # noqa
-from .dde2 import *  # noqa
+from .dde import *  # noqa
 from .opt import *  # noqa
